move GlobalCounter to helpers
break circular import between ops and buffer
