+++ examples/gpt2.py
@@ -4,11 +4,10 @@ import argparse
 from tqdm import trange
 import numpy as np
 import tiktoken
-from tinygrad import Tensor, TinyJit, Device, GlobalCounters
+from tinygrad import Tensor, TinyJit, Device, GlobalCounters, Variable
 from tinygrad.helpers import Timing, DEBUG, getenv, fetch, colored
 from tinygrad.nn import Embedding, Linear, LayerNorm
 from tinygrad.nn.state import torch_load, load_state_dict, get_state_dict
-from tinygrad.shape.symbolic import Variable
 
 MAX_CONTEXT = getenv("MAX_CONTEXT", 128)
 HALF = getenv("HALF")
