+++ test/external/external_benchmark_multitensor_allreduce.py
@@ -1,9 +1,8 @@
 import time
-from tinygrad import Tensor, Device
+from tinygrad import Tensor, Device, GlobalCounters, TinyJit
 from tinygrad.lazy import LazyBuffer
-from tinygrad.ops import ReduceOps, GlobalCounters
+from tinygrad.ops import ReduceOps
 from tinygrad.features.multi import MultiLazyBuffer, all_reduce
-from tinygrad.engine.jit import TinyJit
 from tinygrad.engine.schedule import create_schedule
 from tinygrad.engine.realize import run_schedule
 from tinygrad.helpers import getenv, Context, RING
