+++ tinygrad/__init__.py
@@ -2,5 +2,5 @@ from tinygrad.tensor import Tensor            # noqa: F401
 from tinygrad.engine.jit import TinyJit       # noqa: F401
 from tinygrad.shape.symbolic import Variable  # noqa: F401
 from tinygrad.dtype import dtypes             # noqa: F401
-from tinygrad.ops import GlobalCounters       # noqa: F401
+from tinygrad.helpers import GlobalCounters   # noqa: F401
 from tinygrad.device import Device            # noqa: F401
